Added BepuPhysics dlls compiled in .net 10
Took whatever was on github and just upped .net to 10 and built.

No idea if this was a good idea or if performance is better, but I can always do some testing, at some point. If this even works that is.

We'll see once I can actually render stuff they way I want.